renames in filtsmooth
