+++ src/probnum/diffeq/odesolution.py
@@ -10,7 +10,9 @@ from typing import Optional
 import numpy as np
 
 from probnum import _randomvariablelist, filtsmooth, randvars
-from probnum.filtsmooth.timeseriesposterior import DenseOutputLocationArgType
+from probnum.filtsmooth._timeseriesposterior import (  # Steal a type
+    DenseOutputLocationArgType,
+)
 from probnum.type import FloatArgType, RandomStateArgType, ShapeArgType
