+++ src/probnum/filtsmooth/_timeseriesposterior.py
