+++ src/probnum/filtsmooth/gaussfiltsmooth/_kalmanposterior.py
@@ -12,7 +12,7 @@ from scipy import stats
 from probnum import _randomvariablelist, randvars, statespace, utils
 from probnum.type import FloatArgType, RandomStateArgType, ShapeArgType
 
-from ..timeseriesposterior import DenseOutputLocationArgType, TimeSeriesPosterior
+from .._timeseriesposterior import DenseOutputLocationArgType, TimeSeriesPosterior
 from ._extendedkalman import ContinuousEKFComponent, DiscreteEKFComponent
 from ._unscentedkalman import ContinuousUKFComponent, DiscreteUKFComponent
